fix check for completion

+++ src/pages/api/quest/login.ts
@@ -52,26 +52,31 @@ export default async function handler(
     let user = await getUser(prismaDB, req.body.wallet);
     console.log(user);
     if (user) {
-      const now = new Date();
-      const discordExpiresAt = new Date(
-        Number(user.userDiscordData?.discordExpiresAt!)
-      ); // milliseconds
-      const twitterExpiresAt = new Date(
-        Number(user.userTwitterData?.twitterExpiresAt!)
-      );
-
       if (
-        user.userTwitterData?.twitterExpiresAt === '' ||
-        user.userDiscordData?.discordExpiresAt === '' ||
-        twitterExpiresAt.getTime() < Date.now() ||
-        discordExpiresAt.getTime() < Date.now()
+        !user?.userQuestTask.find(
+          (task) => task.questTask.taskName === 'claim_reward'
+        )
       ) {
-        await prismaDB.userQuestTask.deleteMany({
-          where: {
-            userId: user.id,
-          },
-        });
-        user = await getUser(prismaDB, req.body.wallet);
+        const discordExpiresAt = new Date(
+          Number(user.userDiscordData?.discordExpiresAt!)
+        ); // milliseconds
+        const twitterExpiresAt = new Date(
+          Number(user.userTwitterData?.twitterExpiresAt!)
+        );
+
+        if (
+          user.userTwitterData?.twitterExpiresAt === '' ||
+          user.userDiscordData?.discordExpiresAt === '' ||
+          twitterExpiresAt.getTime() < Date.now() ||
+          discordExpiresAt.getTime() < Date.now()
+        ) {
+          await prismaDB.userQuestTask.deleteMany({
+            where: {
+              userId: user.id,
+            },
+          });
+          user = await getUser(prismaDB, req.body.wallet);
+        }
       }
 
       return res.status(200).json({ data: { message: 'OK', user } });
